requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic
python-dotenv
SQLAlchemy
alembic
psycopg2-binary
pytest
httpx
pytest-asyncio
passlib[bcrypt]
python-jose
cryptography
jinja2
requests
loguru
aiofiles
email-validator
